mine/metadata/requirements.txt
Flask==2.3.3
Pillow==10.0.1
geopy==2.4.0 